+++ src/core/HotelEvent.java
@@ -20,13 +20,13 @@ public abstract class HotelEvent {
 	protected Date checkOut;
 	protected String userEmail;
 	protected ArrayList<Room> rooms;
-	protected double payment;
+	protected int payment;
 	
 	public HotelEvent() {
 		
 	}
 	
-	public HotelEvent(int id, Date checkIn, Date checkOut, String userEmail, ArrayList<Room> rooms, double payment) {
+	public HotelEvent(int id, Date checkIn, Date checkOut, String userEmail, ArrayList<Room> rooms, int payment) {
 		this.id = id;
 		this.checkIn = checkIn;
 		this.checkOut = checkOut;
